requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
drf-spectacular==0.29.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==24.3